cleaned up package.json, organized frontend componeonts
